[TIMOB-18685] Ti.Ui.Window interface should not use HAL types [TIMOB-18716] ListViewAnimationProperties should accept platform-specific properties

+++ Source/TitaniumKit/src/UI/TableView.cpp
@@ -105,7 +105,7 @@ namespace Titanium
 
 		uint32_t TableView::get_sectionCount() const TITANIUM_NOEXCEPT
     	{
-			return sections__.size();
+			return static_cast<uint32_t>(sections__.size());
 		}
 
 		std::string TableView::get_separatorColor() const TITANIUM_NOEXCEPT
